init: move editor from spack/__init__.py to spack.util.editor
